refactor: split github test hotspot
